backend/products/views.py
from django.shortcuts import render
from .serializers import ProductSerializer, ProductImageSerializer, CategorySerializer
from .models import Product, ProductImage, Category
from rest_framework import viewsets
# Create your views here.


class CategoryViewSet(viewsets.ModelViewSet):
    queryset = Category.objects.all()
    serializer_class = CategorySerializer


class ProductViewSet(viewsets.ModelViewSet):
    queryset = Product.objects.all().order_by('-created_at')
    serializer_class = ProductSerializer


class ProductImageViewSet(viewsets.ModelViewSet):
    queryset = ProductImage.objects.all()
    serializer_class = ProductImageSerializer
