backend/requirements.txt
Django>=5.0,<6.0
djangorestframework
django-cors-headers
Pillow
python-dotenv
djangorestframework-simplejwt